后端程序添加功能。

+++ ClientForFront/BackendClient/Model/PoseInfoVm.cs
@@ -12,35 +12,35 @@ namespace BackendClient.Model
 {
     public class PoseInfoVm:ViewModelBase
     {
-        private int _timeStand;
-        public int TimeStand {
+        private double _timeStand;
+        public double TimeStand {
             get => _timeStand;
             set => Set(ref _timeStand,value);
         }
 
-        private int _timeSit=0;
-        public int TimeSit {
+        private double _timeSit=0;
+        public double TimeSit {
             get => _timeSit;
 
             set => Set(ref _timeSit,value);
         }
 
-        private int _timeLie=0;
-        public int TimeLie {
+        private double _timeLie=0;
+        public double TimeLie {
             get => _timeLie;
 
             set => Set(ref _timeLie,value);
         }
 
-        private int _timeDown=0;
-        public int TimeDown {
+        private double _timeDown=0;
+        public double TimeDown {
             get => _timeDown;
 
             set => Set(ref _timeDown,value);
         }
 
-        private int _timeOther=0;
-        public int TimeOther {
+        private double _timeOther=0;
+        public double TimeOther {
             get => _timeOther;
 
             set => Set(ref _timeOther,value);
